+++ libs/core/src/components/icon/icons/cloud-download.ts
@@ -1,0 +1,5 @@
+import { IconCloudDownload } from './cloud-download.component'
+
+IconCloudDownload.define()
+
+export { IconCloudDownload }