+++ libs/core/src/components/icon/icons/pound.ts
@@ -1,0 +1,5 @@
+import { IconPound } from './pound.component'
+
+IconPound.define()
+
+export { IconPound }