+++ libs/core/src/components/textarea/textarea.styles.ts
@@ -1,68 +1,24 @@
 import { css } from 'lit'
 
 export const styles = css`
-  @layer base, simplified;
+  textarea {
+    height: calc(1lh * var(--_lines));
+    min-height: calc(1lh * 4);
+    max-height: calc(1lh * 20);
+    overflow: auto;
+    min-width: 20ch;
+    padding-inline-end: calc(var(--gds-space-xs) + var(--padding-inline-end));
 
-  @layer base {
-    * {
-      box-sizing: border-box;
+    &.resize-manual {
+      resize: vertical;
     }
 
-    :host {
-      display: flex;
-      flex-direction: column;
-      width: 100%;
-      contain: layout;
-      isolation: isolate;
-      gap: var(--gds-space-xs);
+    &.resize-auto {
+      resize: none;
     }
 
-    textarea {
-      appearance: none;
-      background-color: transparent;
-      border: 0;
-      box-sizing: border-box;
-      font-family: inherit;
-      font-size: var(--gds-text-size-detail-m);
-      height: calc(1lh * var(--_lines));
-      line-height: var(--gds-text-line-height-detail-m);
-      margin: unset;
-      min-height: calc(1lh * 4);
-      max-height: calc(1lh * 20);
-      overflow: auto;
-      outline: none;
-      padding: unset;
-      width: 100%;
-      min-width: 20ch;
-      padding-inline-end: calc(var(--gds-space-xs) + var(--padding-inline-end));
-
-      &.resize-manual {
-        resize: vertical;
-      }
-
-      &.resize-auto {
-        resize: none;
-      }
-
-      &.resize-false {
-        resize: none;
-      }
+    &.resize-false {
+      resize: none;
     }
-
-    :host([size='small']) slot[name='lead']::slotted(*) {
-      width: var(--gds-space-m);
-    }
-
-    slot[name='lead']::slotted(*) {
-      min-width: var(--gds-space-l);
-      display: flex;
-    }
-
-    [align-items] ::part(_button) {
-      margin-top: calc(var(--gds-space-2xs) * -1);
-    }
-  }
-
-  @layer simplified {
   }
 `